+++ prov/efa/src/efa_domain.c
@@ -74,7 +74,7 @@ static struct fi_ops_domain efa_ops_domain_rdm = {
 	.size = sizeof(struct fi_ops_domain),
 	.av_open = efa_av_open,
 	.cq_open = efa_rdm_cq_open,
-	.endpoint = rxr_endpoint,
+	.endpoint = efa_rdm_ep_open,
 	.scalable_ep = fi_no_scalable_ep,
 	.cntr_open = efa_cntr_open,
 	.poll_open = fi_poll_create,
